requirements.txt
streamlit>=1.28.1
pandas>=2.2.0
numpy>=1.26.0
pytz>=2023.3
plotly>=5.17.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
